requirements.txt
# Core ML and transformers dependencies
torch>=2.0.0
transformers>=4.35.0
huggingface-hub>=0.30.0,<1.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Progress monitoring
tqdm>=4.65.0

# Configuration and environment
PyYAML>=6.0
python-dotenv>=1.0.0

# Visualization (optional for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# File format support
openpyxl>=3.1.0

# System monitoring
psutil>=5.9.0

# Cloud dependencies (optional - only needed for cloud deployment)
# google-cloud-aiplatform>=1.35.0
# google-cloud-storage>=2.10.0

# Web framework dependencies (optional - only needed for API deployment)
# flask>=2.3.0
# gunicorn>=21.2.0
# uvicorn[standard]>=0.23.0
# fastapi>=0.103.0